Update ApiV1Controller, move tags endpoints to TagsController
